ft_traceroute: rafactor to select
use select instead of iterative loop that blocks, with timeout.
i still need to refactor some stuff to print duration of each probe
for time exceeded messages and print starts correctly if the probe
failed and we didnt get any response or because the probe got dropped
for some reason.

+++ ft_traceroute/ft_traceroute.h
@@ -4,6 +4,7 @@
 #include <argp.h>
 #include <arpa/inet.h>
 #include <bits/posix1_lim.h>
+#include <errno.h>
 #include <net/if.h>
 #include <netdb.h>
 #include <netinet/in.h>
@@ -12,6 +13,7 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <sys/ioctl.h>
 #include <sys/time.h>
 #include <sys/types.h>
@@ -44,7 +46,7 @@ struct opts {
     int current_hop;
 
     long ntransmitted;
-    long ident;
+    uint16_t ident;
 
     struct sockaddr_in source;
     struct sockaddr_in whereto[3];
